+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 1.7.10
+
+- `APIToken`, `APICredential`: added field `refreshToken`.
+
+- `APISecurity`:
+  - Added `onNewAPIToken`, `validateUnknownToken`, `validateRefreshToken`.
+
+- shared_map: ^1.1.9
+- reflection_factory: ^2.4.3
+
 ## 1.7.9
 
 - `EntityHandlerProvider`: